backend/requirements.txt
# DukaPOS Backend - Python 3.11+
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
pydantic-settings==2.6.1
python-multipart==0.0.18
pandas>=2.0,<3
openpyxl==3.1.5
python-escpos==3.1
python-decouple==3.8
bcrypt>=4.0.1,<5

# Packaging
pyinstaller>=6.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
